Copying update and running new manager
