+++ src/services/index.ts
@@ -3,4 +3,4 @@ export * from './get-mod-database';
 export * from './mod-enabler';
 export * from './mod-manager';
 export * from './run-owml';
-export * from './run-self-update';
+export * from './self-update';
